Separate doclink provider into its own file
